pyproject.toml
[tool.poetry]
name = "podcast-pipeline"
version = "0.1.0"
description = "播客音频转录 + 结构化总结流水线"
authors = ["Your Name <your@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
openai-whisper = ">=20240930"
anthropic = "^0.49"
requests = "^2.31"
yt-dlp = "^2024"
torch = {version = "^2.1", source = "pytorch-cu121"}
torchaudio = {version = "^2.1", source = "pytorch-cu121"}
google-genai = ">=1.0"
imageio-ffmpeg = "^0.5"
python-dotenv = "^1.0"

[tool.poetry.scripts]
podcast-pipeline = "podcast_pipeline.cli:main"

[[tool.poetry.source]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
